+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Don't apply default content fields on update (saveItem)
 - Fix GraphQL error when numeric field names are used
 - Fix wysiwyg field (Safari) #24
+- Fix enter on selected result in app search
 
 
 ## 2.2.2 (2022-08-20)